Restrict app access to localhost
The WEB container listens for connection attempts to it on any IP address
available, including any public IPs. In order to prevent accidental exposure of dev
environmnets to the outside world we need to restrict the container to
only forward connections originating from localhost for now. This will
need to be changed before putting the app into production.