Symlink shared files
In the current state it is unclear simply by looking at the file
structure of ./WEB and ./MANAGE what files are required to build the
containers without explicitly reading the Dockerfile. This commit
removes the need to read a Dockerfile by adding a symlink to the
specific files required by each container in their respective
directories. Additionally, to prevent each container's layers from
accidentally pulling in the link, which would point to non-existent
files in the containers, a .dockerignore file has been introduced to
exclude the symlinks.